Add files via upload
It is a machine Learning Project, which is developed to predict the approval of a credit card to a customer with the help of historical data.In this project you can see the working of all classification algorithms. The Python Notebook file provided above has all the code in structured format from preprocessing to Result Analysis.T he PPT_About Project.pptx contains about data set features and visualization for algorithms. Source for data set is Kaggle.